+++ e2e/e2etests/test_bitcoin_std_deposit_and_call_revert_other_address.go
@@ -12,10 +12,6 @@ import (
 )
 
 func TestBitcoinStdMemoDepositAndCallRevertOtherAddress(r *runner.E2ERunner, args []string) {
-	// ARRANGE
-	// Given BTC address
-	r.SetBtcAddress(r.Name, false)
-
 	// Start mining blocks
 	stop := r.MineBlocksIfLocalBitcoin()
 	defer stop()