+++ e2e/e2etests/test_bitcoin_std_memo_inscribed_deposit_and_call.go
@@ -14,10 +14,6 @@ import (
 )
 
 func TestBitcoinStdMemoInscribedDepositAndCall(r *runner.E2ERunner, args []string) {
-	// ARRANGE
-	// Given BTC address
-	r.SetBtcAddress(r.Name, false)
-
 	// Start mining blocks
 	stop := r.MineBlocksIfLocalBitcoin()
 	defer stop()